Refactor resource and scene management to integrate font handling
- Added default font loading and management in ResourceManager, enhancing text rendering capabilities.
- Updated SceneManager and various scenes to utilize the new font system, ensuring consistent text display across the game.
- Removed the obsolete MenuSystem implementation, replacing it with a modular approach for menu handling, improving maintainability and clarity.
- Introduced new menu configuration and navigation handling, streamlining menu interactions and enhancing user experience.
